FnFactory refactor in order to make easy add new types of functions
